py-services/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
numpy==1.21.0
pydantic==1.8.2
python-dotenv==0.19.0